tests/submission/Ms2ger/the-elements-of-html/forms/the-form-element/form-elements-matches.html

<!DOCTYPE html>
<title>form.elements: matches</title>
<link rel="author" title="Ms2ger" href="mailto:ms2ger@gmail.com">
<link rel="help" href="http://www.whatwg.org/html5/#dom-form-elements">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<div id="test">
<form id="form">
<input type="image">
</form>
</div>
<script>
test(function() {
  assert_equals(document.getElementById("form").elements.length, 0);
}, "input type=image should not be present in the form.elements collection")
test(function() {
  var form = document.getElementById("form");
  var i = document.createElement("input");
  i.name = "0";
  form.appendChild(i);
  assert_equals(form.elements[0], i);
  assert_equals(form.elements["0"], i);
  assert_equals(form.elements.namedItem("0"), i);
  assert_array_equals(form.elements, [i]);
  form.removeChild(i);
}, "form.elements should include elements whose name starts with a number");
</script>
